backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
APScheduler==3.10.4
requests==2.31.0
google-genai==1.23.0
pytz==2023.3